final touches

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: hlabouit <hlabouit@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 20:46:13 by hlabouit          #+#    #+#             */
-/*   Updated: 2023/07/11 01:40:46 by hlabouit         ###   ########.fr       */
+/*   Updated: 2023/07/11 16:50:18 by hlabouit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,7 @@ char		**ft_split(char *s, char c);
 char		*ft_strncpy(char *s1, char *s2, int n);
 int			ft_isdigit(int c);
 int			ft_strlen(char *s);
-int			ft_atoi(char *str);
+long		ft_atoi(char *str);
 int			long_negative_nb_error(char **av);
 int			not_integer_error(char **av);
 int			check_input_errors(char **av);
